feat: modify ForumHome Button to move to TextArea
